Replace just::process with an internal process management library
