+++ app/metashell/CMakeLists.txt
@@ -25,6 +25,7 @@ target_link_libraries(metashell
   metashell_core_lib
   metashell_readline_lib
   metashell_data_lib
+  metashell_process_lib
 )
 
 # Wave
