+++ lib/process/CMakeLists.txt
@@ -1,5 +1,5 @@
 # Metashell - Interactive C++ template metaprogramming shell
-# Copyright (C) 2015, Abel Sinkovics (abel@sinkovics.hu)
+# Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -14,5 +14,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-subdirs(data core readline process)
+aux_source_directory(. SOURCES)
+add_library(metashell_process_lib STATIC ${SOURCES})
+
+enable_warnings()
+use_cpp11()
