+++ lib/process/fd_t.hpp
@@ -1,0 +1,38 @@
+#ifndef METASHELL_PROCESS_FD_T_HPP
+#define METASHELL_PROCESS_FD_T_HPP
+
+// Metashell - Interactive C++ template metaprogramming shell
+// Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#ifdef _WIN32
+#include <windows.h>
+#endif
+
+namespace metashell
+{
+  namespace process
+  {
+#ifdef _WIN32
+    typedef HANDLE fd_t;
+#else
+    typedef int fd_t;
+#endif
+
+    fd_t invalid_fd();
+  }
+}
+
+#endif
