+++ lib/process/input_file.hpp
@@ -1,0 +1,45 @@
+#ifndef METASHELL_PROCESS_INPUT_FILE_HPP
+#define METASHELL_PROCESS_INPUT_FILE_HPP
+
+// Metashell - Interactive C++ template metaprogramming shell
+// Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#include "file.hpp"
+
+#include <string>
+
+namespace metashell
+{
+  namespace process
+  {
+    class input_file : public file<input_file>
+    {
+    public:
+      explicit input_file(fd_t fd_);
+
+      size_type read(char* buf_, size_t count_);
+
+      bool eof() const;
+
+      std::string read();
+
+    private:
+      bool _eof;
+    };
+  }
+}
+
+#endif
