+++ test/system/app/core/CMakeLists.txt
@@ -22,6 +22,8 @@ use_cpp11()
 
 target_link_libraries(metashell_core_system_test
   metashell_system_test_lib
+  metashell_data_lib
+  metashell_process_lib
   boost_regex)
 
 # Just
