+++ test/unit/CMakeLists.txt
@@ -23,6 +23,7 @@ use_cpp11()
 target_link_libraries(metashell_unit_test
   metashell_core_lib
   metashell_data_lib
+  metashell_process_lib
 )
 
 # Wave
